requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
django-extensions==3.1.3
django-restframework==0.0.1
djangorestframework==3.12.4
idna==3.2
pycodestyle==2.7.0
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
